OraculoAniversario_Criativo/requirements.txt
gradio
PyPDF2
langchain
langchain-community
langchain-huggingface
gtts
python-dotenv
sentence-transformers
chromadb
